+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 
 ## Version 1.2.0 - dev
 
-### Updates
+### Features
 
 * Added 3 AI difficulty levels.
+
+### Updates
+* Replaced game end modal with general purpose info modal.
+* Before game starts game checks if 2 or more players are set as active.
 * UI tweaks.
 
 ### Bug fixes
@@ -13,6 +17,7 @@
 * Players' starting gold doesn't reset on at the start of the new game.
 * Selected region's info doesn't refresh at the start of the new game.
 * Pausing game can sometimes cause AI player to attack twice.
+* Starting game with all players disabled is breaking the game.
 
 ---